add script to fetch 6 piece tablebases and check paths before downloading tablebases

+++ engines/get_egtb.py
@@ -10,11 +10,13 @@ dtz_links = bs4.BeautifulSoup(requests.get("https://tablebase.lichess.ovh/tables
 for dtz_link in dtz_links:
     link = str(dtz_link).split('href="')[1].split('"')[0]
     if link.endswith(".rtbz"):
-        with open(link, "wb") as file:
-            file.write(requests.get(f"https://tablebase.lichess.ovh/tables/standard/3-4-5-dtz/{link}").content)
+        if os.path.isfile(link) == False or os.path.getsize(link) == 0:
+            with open(link, "wb") as file:
+                file.write(requests.get(f"https://tablebase.lichess.ovh/tables/standard/3-4-5-dtz/{link}").content)
 wdl_links = bs4.BeautifulSoup(requests.get("https://tablebase.lichess.ovh/tables/standard/3-4-5-wdl/").content, 'html.parser').find_all("a")
 for wdl_link in wdl_links:
     link = str(wdl_link).split('href="')[1].split('"')[0]
     if link.endswith(".rtbw"):
-        with open(link, "wb") as file:
-            file.write(requests.get(f"https://tablebase.lichess.ovh/tables/standard/3-4-5-wdl/{link}").content)
+        if os.path.isfile(link) == False or os.path.getsize(link) == 0:
+            with open(link, "wb") as file:
+                file.write(requests.get(f"https://tablebase.lichess.ovh/tables/standard/3-4-5-wdl/{link}").content)
